Final Submission Commit - All tests work with no errors or failures - Game now supports File Input 	- but it does not support splitting since the way I coded my 	splitting did not work the way the file input says 	- and there is not enough time left to change everything main: Added file input to GameController main: minor modifications to HumanPlayer AIDealer and Deck
This Final Submission was made after nearly 35 hours of work put into it since
September 13
- The code is still very VERY inefficient and messy, didn't not have
enough time to fix refactoring of humanPlay (had to make nearly carbon
copy of it to make it work for file input)
- The entire project (in my opinion) is very poorly documented, didn't
not have enough time to carefully comment on everything, much of the
project is left undocumented
- Lessons that I took away from this project are:
	- first and formost design the project carefully after reading
	ALL requirements in full detail
	- start earlier to give enough time for cleaning up the project
	and adding documentation

I personally feel the time, and especially feedback(which was none), we
had on this project was too little, and many of the requirements were
extremely strict. I feel we should've done this project while getting
constant feedback. Then maybe running into issues like these would've
been at the very least, less alarming and significant.

Overall this was a good learning experience and I feel much more ready
for the upcoming group project